reorganize; fix dandi test
